extract logger from lighthouse.
