+++ chrome-launcher/package.json
@@ -19,7 +19,7 @@
     "@types/core-js": "^0.9.41",
     "@types/mkdirp": "^0.3.29",
     "@types/node": "6.0.66",
-    "lighthouse": "2.1"
+    "lighthouse-logger": "^1.0.0"
   },
   "version": "0.1.3",
   "description": "Launch latest Chrome with the devtools-protocol port open",
