+++ lighthouse-core/runner.js
@@ -10,7 +10,7 @@ const GatherRunner = require('./gather/gather-runner');
 const ReportGeneratorV2 = require('./report/v2/report-generator');
 const Audit = require('./audits/audit');
 const emulation = require('./lib/emulation');
-const log = require('./lib/log');
+const log = require('lighthouse-logger');
 const fs = require('fs');
 const path = require('path');
 const URL = require('./lib/url-shim');
