+++ lighthouse-core/test/config/config-test.js
@@ -9,7 +9,7 @@ const Config = require('../../config/config');
 const assert = require('assert');
 const path = require('path');
 const defaultConfig = require('../../config/default.js');
-const log = require('../../lib/log');
+const log = require('lighthouse-logger');
 const Gatherer = require('../../gather/gatherers/gatherer');
 const Audit = require('../../audits/audit');
 const Runner = require('../../runner');
