+++ lighthouse-extension/package.json
@@ -23,5 +23,8 @@
     "package-json-versionify": "^1.0.4",
     "run-sequence": "^1.1.5",
     "through2": "^2.0.1"
+  },
+  "dependencies": {
+    "lighthouse-logger": "^1.0.0"
   }
 }
