+++ lighthouse-logger/README.md
@@ -1,0 +1,4 @@
+# Lighthouse logger
+
+A shared logging utility class for lighthouse and friends.
+
