+++ lighthouse-logger/index.js
