+++ package.json
@@ -76,12 +76,12 @@
   "dependencies": {
     "axe-core": "2.2.1",
     "chrome-devtools-frontend": "1.0.422034",
-    "debug": "2.2.0",
     "devtools-timeline-model": "1.1.6",
     "gl-matrix": "2.3.2",
     "handlebars": "4.0.5",
     "jpeg-js": "0.1.2",
     "json-stringify-safe": "5.0.1",
+    "lighthouse-logger": "^1.0.0",
     "marked": "0.3.6",
     "metaviewport-parser": "0.0.1",
     "mkdirp": "0.5.1",
